# examples/paygate-rust-nuxt/spec/openapi/openapi.yaml
openapi: "3.1.0"
info:
  title: paygate API
  version: "1.0.0"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    paygate is a **payment service**: merchants integrate with it, and it
    integrates ECPay and NewebPay on their behalf as a **platform
    merchant** (ECPay's own term is contracted partner platform merchant). Each merchant keeps its own account at the
    provider; paygate holds the platform credentials that sign for all of them,
    and the money is settled to the merchant's own account, never to paygate's.
    This file is the surface merchants and customers see. What paygate sends the
    providers is in `payment-provider.yaml`; what it sends merchants is in
    `demo-merchant.yaml`.

    ## The steps, and where each credential lives
    | Step | Who calls | Credential |
    |------|-----------|------------|
    | `POST /payments` — create the order AND get the provider's signed form | merchant server | `Authorization: Bearer sk_test_…` |
    | `POST /webhooks/{provider}` — the outcome | the PROVIDER | the provider's signature in the body |
    | the callback to `notify_url` | paygate → merchant | `CheckMacValue` in the body, which the merchant verifies |

    **There is no endpoint here that a customer's browser calls.** The merchant's
    own page renders the form this API returns and submits it to the provider, so
    the customer talks to the merchant and to the provider and never to paygate.
...

...

...

...

...

...

    ## Money
...

...

...
servers:
...
paths:
...
  /dashboard/api-keys:
    post:
      operationId: createApiKey
      summary: Create a secret API key for the caller's merchant
      tags: [Dashboard]
      security: [ { DashboardSession: [] } ]
      description: >
        Generates `sk_test_<merchant slug>_<32 random characters>`, stores its
        SHA-256 and prefix, and returns the raw key exactly once.
      responses:
        "201":
          description: The new key — the only time its raw value is shown
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreatedApiKey" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
...
components:
  securitySchemes:
    MerchantApiKey:
      type: http
      scheme: bearer
      description: A secret key issued in the dashboard.
    ProviderCallbackSignature:
      type: apiKey
      in: query
      name: CheckMacValue
      description: >
        Carried as a FORM FIELD, which OpenAPI has no scheme for, so this
        declaration is documentation. ECPay sends `CheckMacValue` (SHA-256 over
        the sorted parameters wrapped in the HashKey and HashIV the provider
        issued paygate); NewebPay sends `TradeSha` over the encrypted `TradeInfo`.
        paygate recomputes it and compares in constant time before anything is
        parsed or applied.
    DashboardSession:
      type: apiKey
      in: cookie
      name: session
...
  responses:
...
    Unauthenticated:
      description: No valid credential of the kind this endpoint accepts
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    CreatedApiKey:
      type: object
      required: [id, key, key_prefix, created_at]
      properties:
        id: { type: integer }
        key: { type: string }
        key_prefix: { type: string }
        created_at: { type: string, format: date-time }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          enum:
            - INVALID_REQUEST
            - IDEMPOTENCY_KEY_MISSING
            - IDEMPOTENCY_KEY_INVALID
            - IDEMPOTENCY_KEY_IN_USE
            - IDEMPOTENCY_KEY_REUSED
            - MERCHANT_TRADE_NO_TAKEN
            - UNAUTHENTICATED
            - INVALID_CREDENTIALS
            - PAYMENT_NOT_FOUND
            - API_KEY_NOT_FOUND
            - CURRENCY_NOT_SUPPORTED
            - PAYMENT_NOT_REFUNDABLE
            - REFUND_EXCEEDS_REMAINING
            - RANGE_TOO_LARGE
            - RATE_LIMITED
            - PROVIDER_UNAVAILABLE
            - SESSION_STORE_UNAVAILABLE
            - REPORTING_UNAVAILABLE
            - INTERNAL
...